fix(challenge-listing): fix failing test
