+++ src/shared/utils/challenge-detail/helper.jsx
@@ -9,6 +9,7 @@ import { challenge as challengeUtils } from 'topcoder-react-lib';
 import { config } from 'topcoder-react-utils';
 import Prize from 'components/challenge-listing/ChallengeCard/Prize';
 import { BUCKETS, getBuckets } from 'utils/challenge-listing/buckets';
+import { phaseEndDate } from 'utils/challenge-listing/helper';
 
 const Filter = challengeUtils.filter;
 
@@ -30,36 +31,6 @@ export function getChallengeTypeAbbr(track, challengeTypes) {
     return type.abbreviation;
   }
   return null;
-}
-
-/**
-  * Returns phase's end date.
-  * @param {Object} phase
-  * @return {Date}
-  */
-export function phaseEndDate(phase) {
-  // Case 1: phase is still open. take the `scheduledEndDate`
-  // Case 2: phase is not open but `scheduledStartDate` is a future date.
-  // This means phase is not yet started. So take the `scheduledEndDate`
-  if (phase.isOpen || moment(phase.scheduledStartDate).isAfter()) {
-    return new Date(phase.scheduledEndDate);
-  }
-  // for other cases, take the `actualEndDate` as phase is already closed
-  return new Date(phase.actualEndDate);
-}
-
-/**
-  * Returns phase's start date.
-  * @param {Object} phase
-  * @return {Date}
-  */
-export function phaseStartDate(phase) {
-  // Case 1: Phase is not yet started. take the `scheduledStartDate`
-  if (phase.isOpen !== true && moment(phase.scheduledStartDate).isAfter()) {
-    return new Date(phase.scheduledStartDate);
-  }
-  // For all other cases, take the `actualStartDate` as phase is already started
-  return new Date(phase.actualStartDate);
 }
 
 /**
